requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
ultralytics>=8.2.0
opencv-python
pyyaml
numpy
